Selfdriving-Vehicle-V1/requirements.txt
#Linux packages:
#libbluetooth-dev
#libusb-dev
#Python packages:
pybluez
pyusb
nxt-python
pynput
numpy